rag_understanding/requirements.txt
streamlit>=1.28.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.15.0
scikit-learn>=1.0.0
# RAG pipeline deps
openai>=1.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
